### 2019-01-06 #1156 @whedon
[PRE REVIEW]: instrbuilder: a Python package for control of electrical instruments
**Submitting author:** @lucask07 (<a href="http://orcid.org/0000-0002-7236-7202">Lucas J. Koerner</a>)
**Repository:** <a href="https://github.com/lucask07/instrbuilder" target ="_blank">https://github.com/lucask07/instrbuilder</a>
**Version:** 0.1.0
