### 2019-01-07 #1156 @lucask07
@arokem I suggest @casteer (Christopher Steer) who lists software for instrumentation as an interest. It would be best if the reviewer has access to a function generator, oscilloscope or similar to exercise the package.
