### 2019-01-07 ddba80d
update paper.bib due to missing reference in JOSS proof
