+++ paper/paper.bib
@@ -20,3 +20,16 @@
 	Year = {2018},
 	Bdsk-Url-1 = {https://pyvisa.readthedocs.io/en/stable/}
 	}
+
+@article{Arkilic2017,
+	Author = {A. Arkilic and D. B. Allan and T.A. Caswell and L. Li and K. Lauer and S. Abeykoon},
+	Doi = {10.1080/08940886.2017.1289810},
+	Journal = {Synchrotron Radiation News},
+	Month = {Mar.},
+	Number = {2},
+	Pages = {44-45},
+	Publisher = {Informa {UK} Limited},
+	Title = {Towards Integrated Facility-Wide Data Acquisition and Analysis at {NSLS}-{II}},
+	Volume = {30},
+	Year = {2017},
+	Bdsk-Url-1 = {https://doi.org/10.1080/08940886.2017.1289810}}
