### 2019-01-07 #1156 @arokem
Yes. I noticed that too. We'll have chances to do more proof-reading towards the end of the process as well.
