### 2019-01-07 #1156 @arokem
@casteer: would you be willing to review this paper for JOSS?
