### 2019-01-07 #1156 @casteer
Hi, sure looks interesting, we don't have the precise make of oscilloscope mention in the docs but they should be SCPI or VISA, so am happy to accept. The JOSS reviewer's guidelines mentions a checklist of essential features and details, if there is one, please could you post a link? (no worries, if there isn't one, I'm happy to review without it too) Kind regards, Chris.
