### 2019-01-08 #1156 @brainstorm
@arokem Count me in if you need a co-reviewer, I've recently acquired a SCPI/LXI compatible oscilloscope, so perfect timing ;)
