### 2019-01-14 #1172 @whedon
[REVIEW]: instrbuilder: a Python package for control of electrical instruments
**Submitting author:** @lucask07 (<a href="http://orcid.org/0000-0002-7236-7202">Lucas J. Koerner</a>)
**Repository:** <a href="https://github.com/lucask07/instrbuilder" target ="_blank">https://github.com/lucask07/instrbuilder</a>
**Version:** v0.1.9

@brainstorm, @casteer, please carry out your reviews in this issue by updating the checklist below. If you cannot edit the checklist please:
