### 2019-01-16 #1172 @brainstorm
@lucask07 It would help to have some pointers on installing the (generic?) VISA drivers. I own a Rigol MSO5000, [but VISA from NI downloads section](https://www.ni.com/visa/) doesn't lead to any result actually. More hints/pointers would be welcome!

Also, when installing via `pip install instrbuilder`, how is the user supposed to search for the examples you point out? Those small UX details can help saving some time for your users, IMHO.
