### 2019-01-16 #1172 @brainstorm
Thanks much Lucas,

I'm using OSX at the moment.

[From what I can see on your requirements.txt](https://github.com/lucask07/instrbuilder/blob/master/requirements.txt), you are installing a ophyd fork already, right? Are the [installation steps in the README](https://github.com/lucask07/instrbuilder#installation-steps) matching current reality? From the looks of it, a simple `pip install instrbuilder` should suffice after installing the VISA drivers? :)
