### 2019-01-17 lucask07/instrbuilder#21 @lucask07
Ophyd fork is not installed with packaged instrbuilder via pip
The *instrbuilder* [requirements.txt](https://github.com/lucask07/instrbuilder/blob/master/requirements.txt) file lists:

> git+https://github.com/lucask07/ophyd@master#egg=ophyd

But this doesn't work and the master pip version of *ophyd* is installed.

After installation of *instrbuilder* via pip need to run the following pip commands to get the *ophyd* fork:
```
python -m pip uninstall ophyd
python -m pip install git+https://github.com/lucask07/ophyd@master#egg=ophyd
```

To test if the correct *ophyd* fork that supports *Bluesky* and *Instrbuilder* integration:
```python
from ophyd.ee_instruments import generate_ophyd_obj
```
The import will work if the ophyd fork is installed.
