### 2019-01-18 94451ba
add __init__.py to lockin_tests so that it gets packaged
