### 2019-01-18 f025a29
add more URL links to README
