+++ README.md
@@ -8,8 +8,8 @@ https://lucask07.github.io/instrbuilder/build/html/
 
 1. Install a VISA driver (National Instruments provides free downloads). The PyVISA documentation page is an excellent resource to locate a VISA driver for your system: 
 
-* [PyVISA backends](https://pyvisa.readthedocs.io/en/stable/getting.html#backend): an explanation of what is needed.
 * [NI-VISA downloads](https://pyvisa.readthedocs.io/en/stable/getting_nivisa.html#getting-nivisa): links to downloads separated by OS.
+* [PyVISA backends](https://pyvisa.readthedocs.io/en/stable/getting.html#backend): an explanation of what is needed.
 
 2. Install *instrbuilder* using pip.
 
@@ -19,7 +19,7 @@ username$ python -m pip instrbuilder
 
 ### Getting Started 
 
-1. Command lists for Keysight oscilloscopes, function generator, DMM; Rigol DC Power Supply (*commands.csv*) are included in the package at: *instrbuilder/instruments/*.
+1. Command lists for Keysight oscilloscopes, function generator, DMM; Rigol DC Power Supply (*commands.csv*) are included in the package at: [instrbuilder/instruments/](https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/instruments).
 
 Locate the instrument command files using the following python commands:
 
@@ -31,9 +31,9 @@ print('Instrument commmands (csv files) are at: {}'.format(instrument_cmds))
 ```
 Note that you can move the command csv files (and probably should). The location of these needs to be specified in the YAML configuration file.
 
-2. A YAML file is used to track your specific system configurations and instrument addresses (e.g. USB0::0x0957::0x0407::MY44060286::INSTR). Create your system YAML file using the steps below. This will be generated at *~/.instrbuilder/* (where *~* indicates your home directory).
+2. A YAML file is used to track your specific system configurations and instrument addresses (e.g. USB0::0x0957::0x0407::MY44060286::INSTR). Create your system YAML file using the steps below. This will be generated at *~/.instrbuilder/* (where *~* indicates your home directory). See the [example YAML](https://github.com/lucask07/instrbuilder/blob/master/instrbuilder/example_yaml/config.yaml) on GitHub.
 
-3. Try an example in the source code at *instrbuilder/examples/*. A Keysight and Rigol Oscilloscope is demonstrated in: *oscilloscope.py*.
+3. Try an example in the source code at [instrbuilder/examples/](https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/examples). A Keysight and Rigol Oscilloscope is demonstrated in: [oscilloscope.py](https://github.com/lucask07/instrbuilder/blob/master/instrbuilder/examples/oscilloscope.py).
 
 Locate the example directory using the following python commands:
 
@@ -52,10 +52,10 @@ print('Examples are at: {}'.format(examples))
 from instrbuilder import instrument_opening
 instrument_opening.init_yaml(csv_dir = 'where/your/commands_csv/files/are')
 ```
-2. To **add instruments** to the YAML function use the script found in *instrbuilder/examples/add_instruments_to_config.py*. This will prompt the user for information. Note that it may be helpful to have only one instrument connected/powered at a time so that there is no ambiguity:
[…]
