### 2019-01-18 05f8467
update bluesky demos to try/except the ophyd import and print a useful error if it fails
