+++ instrbuilder/bluesky_demo/lockin_tests/freq_res_sr810.py
@@ -16,8 +16,16 @@ from bluesky.callbacks.best_effort import BestEffortCallback
 from bluesky.plans import scan, grid_scan
 from databroker import Broker
 
-from ophyd.device import Kind
-from ophyd.ee_instruments import generate_ophyd_obj
+try:
+    from ophyd.device import Kind
+    from ophyd.ee_instruments import generate_ophyd_obj
+except ModuleNotFoundError:
+    print('Ophyd fork is not installed')
+    print('First, (if needed) uninstall base ophyd:')
+    print('  $ python -m pip uninstall ophyd')
+    print('Next, install the fork:')
+    print('  $ python -m pip install git+https://github.com/lucask07/ophyd@master#egg=ophyd')
+
 from instrbuilder.instrument_opening import open_by_name
 
 RE = RunEngine({})
