+++ instrbuilder/bluesky_demo/lockin_tests/ada2200_spi_tests.py
@@ -21,10 +21,17 @@ from bluesky.callbacks.best_effort import BestEffortCallback
 from bluesky.plans import list_scan
 from databroker import Broker
 
-from ophyd.device import Kind
-from ophyd.ee_instruments import generate_ophyd_obj
-from instrbuilder.instrument_opening import open_by_name
+try:
+	from ophyd.device import Kind
+	from ophyd.ee_instruments import generate_ophyd_obj
+except ModuleNotFoundError:
+	print('Ophyd fork is not installed')
+	print('First, (if needed) uninstall base ophyd:')
+	print('  $ python -m pip uninstall ophyd')
+	print('Next, install the fork:')
+	print('  $ python -m pip install git+https://github.com/lucask07/ophyd@master#egg=ophyd')
 
+from instrbuilder.instrument_opening import open_by_name
 from instrbuilder.instruments import create_ada2200
 
 # ------------------------------------------------

[…]
