### 2019-01-18 lucask07/instrbuilder#21 @lucask07
I believe this is not possible with a PyPi package:
https://github.com/pypa/pip/issues/4187#issuecomment-415067034

Will:
1) remove ophyd from requirements.txt, which eliminates the need for the uninstall step
2) add try/except clauses to ophyd imports which print the instructions to install the ophyd fork
