### 2019-01-18 23a6283
support for Rigol oscilloscope
