+++ requirements.txt
@@ -10,7 +10,6 @@ databroker
 colorama
 matplotlib
 bluesky
-git+https://github.com/lucask07/ophyd@master#egg=ophyd
 pyserial
 wrapt
 PyYAML
