### 2019-01-18 46edd8a
fix config file since that broke Bluesky
