### 2019-01-18 #1172 @lucask07
@brainstorm
I've updated the pip package to version 0.1.5 and improved the README.

The update includes commands.csv, an instrument class, and an example for a Rigol oscilloscope.
Please see a much-updated [README](https://github.com/lucask07/instrbuilder/blob/master/README.md) for more complete install instructions and pointers to the files you asked about above. I tested the Rigol scope example out, it would be great if it works on your end too!

The requirements.txt file was not successfully forcing the ophyd fork installation from GitHub. My research suggests that PyPi does not allow remote URL dependencies.
See the [Ophyd fork install issue](https://github.com/lucask07/instrbuilder/issues/21) that I opened in the instrbuilder repository.
Since ophyd/bluesky is a bonus add on I have:
1. removed ophyd from the requirements.txt
2. added try/except around any ```from ophyd.ee_instruments import ... ``` to let the user know the need to install the ophyd fork for GitHub.
3. Moved the Bluesky/ophyd documentation to the bottom of the README to limit confusion.
