### 2019-01-19 #1172 @brainstorm
Fantastic, thanks much @lucask07 for the detailed README instructions, now I managed to get some good output:

```
$ ./rigol_mso5000_fetch.py
Error:  <class 'ImportError'>
IC (integrated circuit imports failed)
The aardvark.so or dll must be in the cwd or an importable path
Continuing anyways, since many may not use this portion...
Opened Instrument: RIGOL TECHNOLOGIES,MSO5072,SN,00.01.01.02.03
Saving file: rigol_mso5000_screenshot.png
Measured pk-pk voltage of = 0.217338
Measured avg voltage of = -0.03870067
```

Now, other than some polishing on the "safe-to-ignore" Errors (which you raise in the README), I cannot check off the testing part. I know that testing with this type of very specialized hardware is hard (no pun intended), but have you considered [mock testing](https://www.toptal.com/python/an-introduction-to-mocking-in-python) for the instruments for which you know the inputs/outputs of?

A small test (that covers simulated functionality) as a proof of concept should be enough to tick that box off, kudos for this software effort! 👍
