### 2019-01-27 #1172 @casteer
Hi @lucask07, have you got an advice and/or pointers on how to install the NI VISA drivers for linux distributions other the offically-supported ones? There's an reference to this being possible but cumbersome in the [PyVISA Installation docs](https://pyvisa.readthedocs.io/en/stable/getting_nivisa.html#getting-nivisa). I'm using Ubuntu 18.04 and the best I could I find is an unhelpful [note](https://knowledge.ni.com/KnowledgeArticleDetails?id=kA00Z000000PAVZSA4&l=en-GB)...  . I'll see how much I can do though without this in the meantime. Thanks in advance for any help that you're able to give me!
