### 2019-01-27 7d9bf5f
modifications for pytest
