+++ instrbuilder/tests/__init__.py
@@ -1,0 +1,17 @@
+import pytest
+
+'''
+# some module level globals.
+ophyd = None
+reason = ''
+
+try:
+    import ophyd
+except ImportError as ie:
+    # pytestmark = pytest.mark.skip
+    ophyd = None
+    reason = str(ie)
+
+# define a skip condition based on if ophyd is available or not
+requires_ophyd = pytest.mark.skipif(ophyd is None, reason=reason)
+'''
