+++ instrbuilder/instruments.py
@@ -261,6 +261,16 @@ class KeysightMultimeter(SCPI):
 
         return data_array
 
+class TestInstrument(SCPI):
+    def __init__(self,
+                 cmd_list,
+                 comm_handle,
+                 name='tester',
+                 unconnected=True):
+        super().__init__(
+            cmd_list, comm_handle, name=name, unconnected=unconnected)
+
+
 try:
     # for Aardvark SPI control
     from .command import Register
