+++ .travis.yml
@@ -1,0 +1,16 @@
+language: python
+python:
+  - "3.5"
+  - "3.6"
+  - "3.7-dev"  # 3.7 development branch
+# command to install dependencies
+install:
+  - pip install -r requirements.txt
+  - pip install codecov
+  - python setup.py install
+# command to run tests
+script:
+  - coverage run run_tests.py
+  - coverage report -m
+after_success:
+  - codecov
