### 2019-01-27 b3968cb
remove after success travis command
