+++ .travis.yml
@@ -12,5 +12,3 @@ install:
 script:
   - coverage run run_tests.py
   - coverage report -m
-after_success:
-  - codecov
