### 2019-01-27 #1172 @lucask07
@casteer
Apologies, I have not tried to install NI-VISA on an Ubuntu system.
An alternative is the pure Python VISA backend [pyvisa-py](https://github.com/pyvisa/pyvisa-py). This can be installed via `pip`.
If both NI-VISA and pyvisa-py are installed PyVISA defaults to the NI backend. To activate this alternate backend you pass to the VISA resource manager:
```python
visa.ResourceManager('@py')
```
See more information on configuring the backend to PyVISA here: [pyvisa backend info](https://pyvisa.readthedocs.io/en/master/configuring.html

If you don't have NI-VISA installed PyVISA will default to the pyvisa-py backend and this may all work. If you have both installed and you want to use pyvisa-py I will need to make a few simple changes to *instrbuilder*.

Let me know how it goes.
