### 2019-01-27 #1172 @lucask07
@brainstorm
Thank you for the feedback.
For testing, I have added an "unconnected" property to instruments and am using an unconnected instrument to test functionality.

My testing is using the *pytests* package.

The script to run tests is: [run_tests.py](https://github.com/lucask07/instrbuilder/blob/master/run_tests.py)
And the tests are at [tests/test_command.py](https://github.com/lucask07/instrbuilder/blob/master/instrbuilder/tests/test_command.py)

- 8 tests are run, 8 tests pass.
- Code coverage on commands.py is 79%
- Code coverage on scpi.py is 45%

Neither is great, but its a start! and I'm seeing why the tests help development.
