### 2019-01-27 #1172 @lucask07
@arokem
The package has been updated to version 0.1.6 on PyPi and github. I presume we can update the JOSS version to match what is in github (once all changes are done)?
