+++ .travis.yml
@@ -4,11 +4,12 @@ python:
   - "3.6"
   - "3.7-dev"  # 3.7 development branch
 # command to install dependencies
+cache: pip
 install:
   - pip install -r requirements.txt
   - pip install codecov
   - python setup.py install
+  - git clean -xfd
 # command to run tests
 script:
-  - coverage run run_tests.py
-  - coverage report -m
+  - python run_tests.py
