### 2019-01-28 56ec965
update requirements for testing
