+++ requirements.txt
@@ -13,3 +13,5 @@ bluesky
 pyserial
 wrapt
 PyYAML
+pytest
+pytest-cov
