### 2019-01-28 ddf7c50
remove coverage install from travis
