+++ .travis.yml
@@ -1,13 +1,9 @@
 language: python
 python:
-  - "3.5"
   - "3.6"
-  - "3.7-dev"  # 3.7 development branch
 # command to install dependencies
-cache: pip
 install:
   - pip install -r requirements.txt
-  - pip install codecov
   - python setup.py install
   - git clean -xfd
 # command to run tests
