### 2019-02-04 #1172 @arokem
@lucask07 : regarding versions: yes, that's not a problem. The eventual paper will point to a DOI of an archive that you will create (e.g., on zenodo) from the version of the software that incorporates comments/fixes that happened through the review process.
