+++ docs/source/conf.py
@@ -59,7 +59,6 @@ extensions = [
     'IPython.sphinxext.ipython_console_highlighting',
     'matplotlib.sphinxext.plot_directive',
     'sphinx.ext.intersphinx',
-    'sphinxcontrib.ansi',
     'numpydoc']
 
 # Configuration options for plot_directive. See:
@@ -103,8 +102,8 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path .
+# exclude_patterns = ['tutorial.rst']
 exclude_patterns = []
-
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = 'sphinx'
 
