+++ docs/source/installation.rst
@@ -2,79 +2,86 @@ Installation
 ******************************
 1. Install a VISA driver (National Instruments provides free downloads). The PyVISA documentation page is an excellent resource to locate a VISA driver for your system: 
 
-* [NI-VISA downloads](https://pyvisa.readthedocs.io/en/stable/getting_nivisa.html#getting-nivisa): links to downloads separated by OS.
-* [PyVISA backends](https://pyvisa.readthedocs.io/en/stable/getting.html#backend): an explanation of what is needed.
+* `NI-VISA downloads <https://pyvisa.readthedocs.io/en/stable/getting_nivisa.html#getting-nivisa>`_ links to downloads separated by OS.
+* `PyVISA backends <https://pyvisa.readthedocs.io/en/stable/getting.html#backend>`_: an explanation of what is needed.
 
 2. Install *instrbuilder* using pip.
 
-```console
-username$ python -m pip instrbuilder 
-```
+.. code-block:: console
+
+	username$ python -m pip instrbuilder 
+
 
 Getting Started 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-1. Command lists for Keysight oscilloscopes, function generator, DMM; Rigol DC Power Supply (*commands.csv*) are included in the package at: [instrbuilder/instruments/](https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/instruments).
+1. Command lists for Keysight oscilloscopes, function generator, DMM; Rigol DC Power Supply (*commands.csv*) are included in the package at: `instrbuilder/instruments/ <https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/instruments>`_
 
-Locate the instrument command files using the following python commands:
+Locate the instrument command files on your system using the following python commands:
 
-```python
-import instrbuilder
-init_file_loc = instrbuilder.__file__
-instrument_cmds = init_file_loc.replace('__init__.py', 'instruments/')
-print('Instrument commmands (csv files) are at: {}'.format(instrument_cmds))
-```
+::
+
+	import instrbuilder
+	init_file_loc = instrbuilder.__file__
+	instrument_cmds = init_file_loc.replace('__init__.py', 'instruments/')
+	print('Instrument commmands (csv files) are at: {}'.format(instrument_cmds))
+
 Note that you can move the command csv files (and probably should). The location of these needs to be specified in the YAML configuration file.
 
-2. A YAML file is used to track your specific system configurations and instrument addresses (e.g. USB0::0x0957::0x0407::MY44060286::INSTR). Create your system YAML file using the steps below. This will be generated at *~/.instrbuilder/* (where *~* indicates your home directory). See the [example YAML](https://github.com/lucask07/instrbuilder/blob/master/instrbuilder/example_yaml/config.yaml) on GitHub.
+2. A YAML file is used to track your specific system configurations and instrument addresses (e.g. USB0::0x0957::0x0407::MY44060286::INSTR). Create your system YAML file using the steps below. This will be generated at *~/.instrbuilder/* (where *~* indicates your home directory). See the `example YAML <https://github.com/lucask07/instrbuilder/blob/master/instrbuilder/example_yaml/config.yaml>`_ on GitHub.
 
[…]

[…]
