### 2019-02-07 d66d34b
address issue #22
