+++ docs/source/contributing.rst
@@ -1,0 +1,21 @@
+Contributing and Issues
+**************************************************
+Contributors are very welcome! 
+
+Navigate to the GitHub repo at:
+
+`instrbuilder <https://github.com/lucask07/instrbuilder>`_
+
+You could: 
+
+* Add python code 
+* Add a new interface 
+* Contribute a new instrument commands.csv file 
+
+====================
+Issues
+====================
+
+File issues at:
+
+`Issues <https://github.com/lucask07/instrbuilder/issues>`_
