### 2019-02-07 da9de13
fix contributing help page
