+++ docs/source/contributing.rst
@@ -5,6 +5,7 @@ Contributors are very welcome!
 Navigate to the GitHub repo at: `instrbuilder <https://github.com/lucask07/instrbuilder>`_
 
 You could: 
+
 * Add python code 
 * Add a new interface 
 * Contribute a new instrument commands.csv file 
