### 2019-02-08 879164b
remove relative imports and build docs with Pyvisa-py howto
