### 2019-02-08 754d2d6
documentation bug fixes
