+++ docs/source/linux_visa.rst
@@ -20,7 +20,10 @@ Check VISA status:
 Ubuntu requires "rules" to be added for a user to fully access the USB device. 
 
 Add the following line to /etc/udev/rules.d/99-com.rules:
-SUBSYSTEM=="usb", MODE="0666", GROUP="usbusers"
+
+.. code-block:: console
+
+	SUBSYSTEM=="usb", MODE="0666", GROUP="usbusers"
 
 Create the 99-com.rules file if necessary.
 
@@ -42,3 +45,7 @@ Finally check to see if you can detect a USB connected instrument using PyVISA:
 	address = rm.list_resources()
 
 If the Python code returns a device address you are now good to go!
+
+
+For more info regarding USB permissions and the udev rules see `this stackoverflow question <https://stackoverflow.com/questions/52256123/unable-to-get-full-visa-address-that-includes-the-serial-number>`_
+
