+++ docs/build/html/_sources/registers.rst.txt
@@ -5,5 +5,6 @@ A Class is available for organizing, reading, and writing registers as would be 
 
 .. module:: command
 
-.. py:class:: Register    
+.. autoclass:: Register     
+   :members: 
 
