### 2019-02-08 #1172 @lucask07
Hi @casteer,
I researched and tested a way to get this up and running with Ubuntu. With Ubuntu, we can use the pure Python VISA back end. This works on my system using a function generator.
See the detailed setup instructions which I've added to the instrbuilder documentation:
[Visa with Linux](https://lucask07.github.io/instrbuilder/build/html/linux_visa.html)
