### 2019-02-17 #1172 @lucask07
Hi @casteer,
Any luck or any questions with the Linux setup?
