### 2019-02-19 #1172 @dancekickrun
Hi Lucas, I'm sorry for the delay and the new installation page has greatly helped, thank you. I've got it installed now with no errors and will be testing with one of our instruments over the next couple of days. Thanks again for your help. Best wishes.
