### 2019-02-19 #1172 @casteer
Hi, Thanks so much @lucask07 for the new help page and I'm sorry for the delay. I've got it install and will be instrument testing over the next couple of days. Best wishes.
