### 2019-02-24 ea49800
label each analysis code with paper figure number
