+++ instrbuilder/bluesky_demo/lockin_analysis/snr_sr810_analysis.py
@@ -1,4 +1,12 @@
+# Lucas J. Koerner
+# 05/2018
+# koerner.lucas@stthomas.edu
+# University of St. Thomas
 
+'''
+analysis of SR810 signal to noise ratio
+paper Figure 5
+'''
 # standard library imports
 import os
 
@@ -9,11 +17,10 @@ from matplotlib import rcParams
 from databroker import Broker
 
 from plot_configs import params, dpi, figure_dir
-
 rcParams.update(params)
-rcParams.update(params)
+SAVE_FIGS = False
 
-db = Broker.named('local_file')  # a broker poses queries for saved data sets)
+db = Broker.named('local_file')  # a broker poses queries for saved data sets
 
 # get data into a pandas data-frame
 uid = '6c22a3cf-5851-49c8-9086-77f759701284'
@@ -35,7 +42,8 @@ ax2.plot(df['att_val'], df['lockin_read_buffer_std']*1000, marker='x', color='k'
 ax2.set(xlabel='Attenuation. [dB]', ylabel='Noise [mV RMS]')
 plt.grid(True)
 plt.tight_layout()
-plt.savefig(os.path.join(figure_dir, 'snr_sr810.eps'))
+if SAVE_FIGS:
+	plt.savefig(os.path.join(figure_dir, 'snr_sr810.eps'))
 
 snr = df['lockin_read_buffer_mean']/df['lockin_read_buffer_std']
 snr_bits = np.log2(snr)
@@ -54,4 +62,5 @@ ax2.plot(df['att_val'], df['lockin_read_buffer_std']*1000, marker='x', color='k'
 ax2.set(xlabel='Attenuation. [dB]', ylabel='Noise [mV RMS]')
 plt.grid(True)
 plt.tight_layout()
-plt.savefig(os.path.join(figure_dir, 'snr_sr810_bits.eps'))
+if SAVE_FIGS:
+	plt.savefig(os.path.join(figure_dir, 'snr_sr810_bits.eps'))

[…]
