### 2019-02-25 d4f2342
Added metadata parsing function so that all experiments print out metadata
