### 2019-02-26 8666e33
add README to analysis code that points to figshare data repo
