### 2019-02-26 1ff649c
change directory in YAML snippet of readme
