### 2019-02-28 b9eda6e
fit phase tf of ada2200 to extract max, min, zero cross
