+++ instrbuilder/bluesky_demo/lockin_analysis/phase_res_ada2200_analysis.py
@@ -12,6 +12,7 @@ import os
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import rcParams
+import scipy.interpolate as inter
 from databroker import Broker
 
 from plot_configs import params, dpi, figure_dir
@@ -65,8 +66,44 @@ plt.xlabel('Phase [deg]')
 plt.ylabel('Magnitude [V]')
 plt.grid(True)
 
-
 if SAVE_FIGS:
 	plt.savefig(os.path.join(figure_dir, 'phase_res_ada2200.eps'))
 plt.show()
 
+# create a fit to the data to extrapolate:
+	#1) the phase of the zero-crossing 
+	#2) the conversion gain at the peak 
+x = df['osc_meas_phase']
+x_sort = np.sort(df['osc_meas_phase'])
+x_idx = np.argsort(df['osc_meas_phase'])
+x_fit = np.linspace(x_sort[0], x_sort[-1], 100000)
+data = np.array(df['dmm_burst_volt_timer_mean'] - offset)
+s_fit = inter.UnivariateSpline (x_sort, data[x_idx], s=0.001)
+
+PLOT_FIT = True
+if PLOT_FIT:
+	plt.plot(x_fit, s_fit(x_fit), 'r')
+
+# use fit to find phase of zero-crossing 
+#	find sign change and positive going
+print('-'*60)
+print('ADA2200 phase analysis to compare to datasheet figure 15')
+sign_fit = np.sign(s_fit(x_fit))
+sign_change = sign_fit[1:]*sign_fit[0:-1]
+
+pos_zerocross = np.intersect1d(np.where(sign_fit < 0), np.where(sign_change == -1))
+print('Pos zero cross = {}'.format(x_fit[pos_zerocross[0]]))
+
+# use fit to find max and min voltage values and determine conversion value
+
+# get function generator amplitude from baseline measurements (pk-pk)
+fgen_v = header.table('baseline')['fgen_v'].iloc[0] 
+
+maxval = np.max(s_fit(x_fit))
+minval = np.min(s_fit(x_fit))
+print('Max value = {}'.format(np.max(maxval)))
+print('Min value = {}'.format(np.max(minval)))
+
+print('Conversion gain from max = {}'.format(maxval/(2/(fgen_v*np.sqrt(2)))))
+print('Conversion gain from min = {}'.format(minval/(2/(fgen_v*np.sqrt(2)))))
+print('-'*60)
