+++ instrbuilder/bluesky_demo/lockin_analysis/README.md
@@ -7,10 +7,17 @@ Once publised the dataset will have a DOI of: <10.6084/m9.figshare.7768352>
 
 The analysis requires databroker, matplotlib, scipy and numpy. 
 
+the matplotlib figures require latex and dvipng. Installation on Ubuntu:
+
+```terminal 
+sudo apt-get install texlive-xetex
+sudo apt-get install dvipng
+```
+
 To setup a databroker configuration to point to the directory that stores the data refer to:
 [Data broker configurations](http://nsls-ii.github.io/databroker/configuration.html)
 
-The database configuration file should be named *local_file.yml* to match the analysis scripts. The file contents should be:
+The database configuration file should be named *local_file.yml* to match the analysis scripts. The file contents should be and placed within *~/.config/databroker/:
 
 ```yaml
 description: 'lightweight personal database'
@@ -18,7 +25,7 @@ metadatastore:
     module: 'databroker.headersource.sqlite'
     class: 'MDS'
     config:
-        directory: 'data_directory'
+        directory: your_directory
         timezone: 'US/Eastern'
 assets:
     module: 'databroker.assets.sqlite'
@@ -28,5 +35,4 @@ assets:
 ```
 
 The script ```run_all_analysis.py``` creates each figure for the paper under review at IEEE TIM. 
-
 
