+++ instrbuilder/bluesky_demo/lockin_analysis/README.md
@@ -1,7 +1,7 @@
 The analysis code for lock-in amplifier analysis run with Bluesky and instrbuilder is found in this folder. 
 
 The data is available via figshare at: 
-[data repository link](https://figshare.com/s/ce1554937c4adc2f1de2) the dataset has a DOI of: <doi:10.6084/m9.figshare.7768352>
+[data repository link](https://figshare.com/s/ce1554937c4adc2f1de2) the dataset has a DOI of: <https://doi.org/10.6084/m9.figshare.7768352>
 
 The analysis requires databroker, matplotlib, scipy and numpy and the matplotlib figures require latex and dvipng. Matplotlib helpers installation on Ubuntu:
 
