+++ instrbuilder/bluesky_demo/lockin_analysis/README.md
@@ -2,17 +2,14 @@ The analysis code for lock-in amplifier analysis run with Bluesky and instrbuild
 
 The data is available via figshare at: <https://doi.org/10.6084/m9.figshare.7768352>
 
-The analysis requires databroker, matplotlib, scipy and numpy and the matplotlib figures require latex and dvipng. Matplotlib helpers installation on Ubuntu:
+The analysis requires databroker, matplotlib, scipy and numpy. The matplotlib figures require latex and dvipng. Matplotlib helpers installation on Ubuntu:
 
 ```terminal 
 sudo apt-get install texlive-xetex
 sudo apt-get install dvipng
 ```
 
-To setup a databroker configuration to point to the directory that stores the data refer to:
-[Data broker configurations](http://nsls-ii.github.io/databroker/configuration.html)
-
-The database configuration file should be named *local_file.yml* to match the analysis scripts. The file contents should be and placed within *~/.config/databroker/:
+Databroker requires a configuration file to locate the data. The data configuration file should be named *local_file.yml* to match the analysis scripts and placed within *~/.config/databroker/ 
 
 ```yaml
 description: 'lightweight personal database'
@@ -31,3 +28,7 @@ assets:
 
 The script ```run_all_analysis.py``` creates each figure for the paper under review at IEEE TIM. 
 
+The system file open limit may need to be modified:
+```terminal
+ulimit -Sn 4095
+```
