+++ instrbuilder/bluesky_demo/lockin_analysis/README.md
@@ -1,4 +1,4 @@
-The analysis code for lock-in amplifier analysis run with Bluesky and instrbuilder is found in this folder. 
+The analysis code for lock-in amplifier experiments executed with Bluesky and instrbuilder is found in this folder. 
 
 The data is available via figshare at: <https://doi.org/10.6084/m9.figshare.7768352>
 
