### 2019-03-02 09ec985
new metadata processing code and example
