+++ instrbuilder/bluesky_utils/metadata.py
@@ -25,84 +25,89 @@ def equivalence_check(config1, config2):
     """
     verbose = False
     debug = False
-    print('------------')
-    for k1, k2 in zip(config1.keys(), config2.keys()):
+    print('Located {} configuration keys to compare:'.format(len(list(config1.keys()))))
+    for k1 in config1.keys():
+        if k1 not in config2.keys():
+            print('Skipping {} not in config2'.format(k1))
         if verbose:
             print('Checking {}'.format(k1))
-        if config1[k1]['data'][k1] != config2[k2]['data'][k2]:
-            print('Warning!! Found configuration difference:')
-            print('  Data-set 1: {} = {}'.format(k1, config1[k1]['data'][k1]))
-            print('  Data-set 2: {} = {}'.format(k2, config2[k2]['data'][k2]))
-        if debug:
-            print('  Data-set 1: {} = {}'.format(k1, config1[k1]['data'][k1]))
-            print('  Data-set 2: {} = {}'.format(k2, config2[k2]['data'][k2]))
+        if config1[k1]['data'][k1] != config2[k1]['data'][k1]:
+            print(' Found configuration difference:')
+            print('  Run 1: {} = {}'.format(k1, config1[k1]['data'][k1]))
+            print('  Run 2: {} = {}'.format(k1, config2[k1]['data'][k1]))
+        elif debug:
+            print('  Run 1: {} = {}'.format(k1, config1[k1]['data'][k1]))
+            print('  Run 2: {} = {}'.format(k1, config2[k1]['data'][k1]))
 
-list_configs()
-db = Broker.named('local_file')
-# this returns the data directory
-print('Databroker config:')
-pprint.pprint(db.get_config())
-# --------------------------------------------------------------------
-#                  Summarize the last 50 experiments
-# --------------------------------------------------------------------
-headers = db[-50:]  # get the last 50 headers (experiment metadata)
-df = pd.DataFrame.from_dict([h.start for h in headers])  # put the run start metadata into a pandas dataframe
+if __name__ == '__main__':
+    list_configs()
+    db = Broker.named('local_file')
+    # this returns the data directory
+    print('Databroker config:')
+    pprint.pprint(db.get_config())
+    # --------------------------------------------------------------------
+    #                  Summarize the last 50 experiments
+    # --------------------------------------------------------------------
+    headers = db[-50:]  # get the last 50 headers (experiment metadata)
+    df = pd.DataFrame.from_dict([h.start for h in headers])  # put the run start metadata into a pandas dataframe
 
-# Create a human-readable time index
-df['human_time'] = df.apply(lambda row: datetime.fromtimestamp(row['time']).strftime('%Y-%m-%d %H:%M:%S'), axis=1)
-cols = df.columns.tolist()
+    # Create a human-readable time index
+    df['human_time'] = df.apply(lambda row: datetime.fromtimestamp(row['time']).strftime('%Y-%m-%d %H:%M:%S'), axis=1)
+    cols = df.columns.tolist()
 
-# re-order the dataframe so that uid is the leftmost column
-new_cols = [f for f in cols if f != 'uid']
-new_cols.insert(0, 'uid')
+    # re-order the dataframe so that uid is the leftmost column
+    new_cols = [f for f in cols if f != 'uid']
+    new_cols.insert(0, 'uid')
 
-# print out the header information with a few useful columns
-df_short = df[['uid', 'human_time', 'purpose', 'plan_name', 'dut', 'num_points']]
-print(tabulate(df_short, headers='keys', tablefmt='psql'))
+    # print out the header information with a few useful columns
+    df_short = df[['uid', 'human_time', 'purpose', 'plan_name', 'dut', 'num_points']]
[…]
