### 2019-03-07 9dd442c
update setup to version 1.8
