### 2019-04-08 #1172 @arokem
@lucask07 : I apologize for the delay here. As @brainstorm has given this the 👍 and @casteer seems to have disappeared, I am inclined to move ahead with your article. I have the following comments:

1. The authors in the bibliography sometimes don't make sense. For PyVISA, please ask the maintainers how to refer to the software (maybe they should publish in JOSS...). For the SCPI syntax and style guide, I believe that no author should be named.

2. We might need a reference for "Aardvark controller" (e.g., who manufactures this and where?).

3. Is there already a reference for the "in press" article mentioned in the last paragraph?

If you could address these comments, we can move ahead with the process of accepting your article for publication. Thanks for your patience!
