### 2019-04-10 10d0001
add metadata examples and instrbuilder joss bibtex
