+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 *instrbuilder* is an open-source Python package for control of electrical instruments. This package eases the development of high-level "drivers" to interface with oscilloscopes, power supplies, function generators, multimeters, or any instrument that uses ASCII textual string communication (such as SCPI [@scpi1999standard]). *instrbuilder* is particularly suited for prototyping and automating experiments in research laboratory setups within an IPython terminal. Our goals are to accelerate the development of automated data collection and improve the reproducibility of laboratory experiments.
 
 # Summary
-*instrbuilder* reads and writes data from instruments with serial (using PySerial [@pyserial]) or USB (using PyVISA [@pyvisa]) interfaces and leverages the consistency of SCPI commands so that new instrument "drivers" can be created with a simple CSV file (example command files are included in the package). 
+*instrbuilder* reads and writes data from instruments over a serial interface by using PySerial [@pyserial] or over a USB interface by using PyVISA [@pyvisa]. *instrbuilder* leverages the consistency of SCPI commands so that new instrument "drivers" can be created with a simple CSV file (example command files are included in the package). 
 An example write command to configure the offset of a function generator:
 ```python
 fg.set('offset', 0.5)
@@ -34,11 +34,11 @@ offset = fg.get('offset')
 ```
 Nonstandard instrument commands may also be generated by writing Python code. Each instrument object is comprised of a communication handle and a list of commands. Each command may be automatically tested for communication errors or unexpected return values and stores a help string. To support experimental reproducibility, commands that access a configuration state are specified so that a complete archive of the instrument configuration may be automatically read and logged. Instrument addresses and other environment properties may be stored in a YAML system configuration file so that the user may name each instrument and simply open instruments by name. *instrbuilder* prompts the user through the steps to generate the YAML system configuration file. 
 
-Documentation is online https://lucask07.github.io/instrbuilder/build/html/ and [examples](https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/examples) are available in the [GitHub repository](https://github.com/lucask07/instrbuilder). These minimal working examples (MWEs) include control of a function generator, power supply, oscilloscope, lock-in amplifier, and an integrated circuit (IC) interfaced with an Aardvark controller. 
+Documentation is online https://lucask07.github.io/instrbuilder/build/html/ and [examples](https://github.com/lucask07/instrbuilder/tree/master/instrbuilder/examples) are available in the [GitHub repository](https://github.com/lucask07/instrbuilder). These minimal working examples (MWEs) include control of a function generator, power supply, oscilloscope, lock-in amplifier, and an integrated circuit (IC) interfaced with an Aardvark I2C/SPI controller (Total Phase, Santa Clara, CA). 
 
 *instrbuilder* supports a Register class that targets registers of an IC. The IC class objects include a dictionary of the registers and have methods for reading and writing these registers via SPI or I2C. 
 
-*instrbuilder* is part of a Python software suite that uses Bluesky [@Arkilic2017] and associated modules created for x-ray science by a team at the NSLS-II (see https://nsls-ii.github.io/) to generate an instrument and experiment control environment that facilitates reproducible experiments. A full-length manuscript that describes and demonstrates the use of *instrbuilder* and *Bluesky* is in press. *instrbuilder* is available at (https://pypi.org/project/instrbuilder/) and can be installed with pip. 
[…]
