+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @manual{scpi1999standard,
-	Author = {SCPI Consortium and others},
+	Author = {{SCPI Consortium}},
 	Date-Modified = {2018-10-31 23:05:22 -0500},
 	Journal = {USA, May},
 	Title = {Standard Commands for Programmable Instruments ({SCPI}), Volume 1: Syntax and Style},
@@ -14,7 +14,7 @@
 	}
 
 @misc{pyvisa,
-	Author = {PyVISA Authors},
+	Author = {{PyVISA Authors}},
 	Title = {{P}y{VISA} Documentation},
 	Url = {https://pyvisa.readthedocs.io/en/stable/},
 	Year = {2018},
