### 2019-04-10 #1172 @lucask07
@arokem : Thank you for your comments.

1. I've updated the PyVISA author entry in the Bibtex so that it is not abbreviated. It is sensible now. I've also added an issue to the PyVISA repository to see if there is a preferred citation. For the SCPI style guide, I also fixed the abbreviation. However, the author is not empty since that would break the (author, year) citation format in the text.

2. The text now reads Aardvark controller (Total Phase, Santa Clara, CA).

3. There is not yet a citation available for this article. Instead, I've added a link to a MWE that uses both instrbuilder and the NSLS-II Bluesky suite.

4. Small grammatical changes and modifications to eliminate double parantheses (due to the citations).

Let me know if this seems OK.
