### 2019-04-10 ae4d0b7
switch to short urls in bibtex
