+++ paper/paper.bib
@@ -8,17 +8,15 @@
 @misc{pyserial,
 	Author = {Chris Liechti},
 	Title = {py{S}erial Documentation},
-	Url = {http://pyserial.readthedocs.io/en/latest/},
-	Year = {2017},
-	Bdsk-Url-1 = {http://pyserial.readthedocs.io/en/latest/}
+	Url = {http://pyserial.readthedocs.io},
+	Year = {2017}
 	}
 
 @misc{pyvisa,
 	Author = {{PyVISA Authors}},
 	Title = {{P}y{VISA} Documentation},
-	Url = {https://pyvisa.readthedocs.io/en/stable/},
-	Year = {2018},
-	Bdsk-Url-1 = {https://pyvisa.readthedocs.io/en/stable/}
+	Url = {https://pyvisa.readthedocs.io},
+	Year = {2018}
 	}
 
 @article{Arkilic2017,
