### 2019-04-10 #1172 @lucask07
In response to https://github.com/pyvisa/pyvisa/issues/407 switched to short links for both PyVISA and PySerial.
