### 2019-04-14 #1172 @arokem
This all looks good to me. Could you please create an archive of the software as it is now (e.g., using [zenodo](https://zenodo.org)) and post the DOI here?
