### 2019-04-15 #1172 @lucask07
@arokem
The version on GitHub and PyPi is now v0.1.9
