### 2019-04-15 #1172 @arokem
👋 @openjournals/joss-eics this submission is ready to be accepted
